+++ src/ovb_common/rtp/packet.h
@@ -6,7 +6,7 @@ class RTPPacket
 {
 public:
 	RTPPacket(uint8_t InPType, uint16_t InFrameNB, uint32_t InTime, const uint8_t* InData, size_t InDataLength, bool InbIsLast);
-	RTPPacket(const uint8_t* InData, int InDataLength);
+	RTPPacket(const uint8_t* InData, int64_t InDataLength);
 
 	uint8_t* GetHeader();
 	size_t	 GetHeaderSize();
